backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
browser-use>=0.1.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
anthropic>=0.39.0
langchain-anthropic>=0.3.0
langchain>=0.3.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
langchain-google-genai>=4.0.0
stagehand>=3.0.0
